Add MixinCreateWorldScreen to add Paxi packs to the datapack selection screen
In the client, is possible for the datapack selection screen to be loaded before the server is initialized, so the hook that loads Paxi packs hasn't run yet. It does run *after* the datapack selection screen is loaded, but this results in Paxi packs getting added in the wrong order, which can cause them to be overridden by packs added by Fabric mods.

This forces the Paxi packs to be registered before the datapack selection screen is displayed.

Only for Fabric at the moment.